chore(EMS-1523): split up API email functions
