+++ src/api/emails/password-reset-link/index.ts
@@ -1,0 +1,29 @@
+import { EMAIL_TEMPLATE_IDS } from '../../constants';
+import { callNotify } from '../call-notify';
+import { EmailResponse } from '../../types';
+
+/**
+ * passwordResetLink
+ * Send "reset your password" email to an account
+ * @param {String} Email address
+ * @param {String} First name
+ * @param {String} Password reet token
+ * @returns {Object} callNotify response
+ */
+export const passwordResetLink = async (urlOrigin: string, emailAddress: string, name: string, passwordResetHash: string): Promise<EmailResponse> => {
+  try {
+    console.info('Sending email for account password reset');
+
+    const templateId = EMAIL_TEMPLATE_IDS.ACCOUNT.PASSWORD_RESET;
+
+    const variables = { urlOrigin, name, passwordResetToken: passwordResetHash };
+
+    const response = await callNotify(templateId, emailAddress, variables);
+
+    return response;
+  } catch (err) {
+    console.error(err);
+
+    throw new Error(`Sending email for account password reset ${err}`);
+  }
+};
